move : Sell 도메인 Test 분리
