+++ src/test/java/bc1/gream/domain/sell/repository/SellRepositoryTest.java
@@ -1,12 +1,12 @@
-package bc1.gream.domain.order.repository;
+package bc1.gream.domain.sell.repository;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bc1.gream.domain.gifticon.repository.GifticonRepository;
 import bc1.gream.domain.order.entity.Gifticon;
-import bc1.gream.domain.order.entity.Sell;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
+import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.user.entity.User;
 import bc1.gream.domain.user.repository.UserRepository;
 import bc1.gream.global.common.ResultCase;
