+++ src/test/java/bc1/gream/domain/sell/service/SellNowServiceIntegrationTest.java
@@ -1,11 +1,11 @@
-package bc1.gream.domain.order.service;
+package bc1.gream.domain.sell.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.buy.repository.BuyRepository;
-import bc1.gream.domain.order.dto.request.SellNowRequestDto;
-import bc1.gream.domain.order.dto.response.SellNowResponseDto;
+import bc1.gream.domain.sell.dto.request.SellNowRequestDto;
+import bc1.gream.domain.sell.dto.response.SellNowResponseDto;
 import bc1.gream.test.BaseIntegrationTest;
 import bc1.gream.test.BuyTest;
 import bc1.gream.test.CouponTest;
