+++ src/test/java/bc1/gream/test/BaseIntegrationTest.java
@@ -3,9 +3,9 @@ package bc1.gream.test;
 import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.gifticon.repository.GifticonRepository;
 import bc1.gream.domain.order.repository.OrderRepository;
-import bc1.gream.domain.order.repository.SellRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
+import bc1.gream.domain.sell.repository.SellRepository;
 import bc1.gream.domain.user.entity.Coupon;
 import bc1.gream.domain.user.entity.User;
 import bc1.gream.domain.user.repository.CouponRepository;
